Attempt to configure magithub.
This adds the auth-sources variable to my config to ensure that the
right file is being pulled with my github token. However it appears that
ghub does not pull from this file as I'm getting the wizard still even
though I'm not supposed to cause I have 2FA turned on. I've opened an
issue magit/ghub#57 so hopefully they'll get back to me.